requirements.txt
boto3
google-cloud-storage
pytest
pytest-mock
moto
mock